Refactoring
1. 함수들을 계산기 코드에 옮김
2. 각 다형성 별로 다른 계산결과를 반환하도록 상속된 계산기 클래스를 생성(TradegyCalculator.java, ComedyCalculator.java)
3. 타입에 맞는 계산기를 반환하기위한 CalculatorFactory.java를 만듦
